adding model

+++ src/build_model_GENERIC_multiscale_auto.py
@@ -284,9 +284,9 @@ def main():
     train_predictors = torch.tensor(training_predictors).float().to(device)
     val_predictors = torch.tensor(validation_predictors).float().to(device)
     test_predictors = torch.tensor(test_predictors).float().to(device)
-    train_target = torch.tensor(training_target, dtype=torch.long).to(device)
-    val_target = torch.tensor(validation_target, dtype=torch.long).to(device)
-    test_target = torch.tensor(test_target, dtype=torch.long).to(device)
+    train_target = torch.tensor(training_target).float().to(device)
+    val_target = torch.tensor(validation_target).float().to(device)
+    test_target = torch.tensor(test_target).float().to(device)
 
     #build dataset from tensors
     tr_dataset = utils.TensorDataset(train_predictors,train_target) # create your datset
